+++ js/jquery.terminal-src.js
@@ -1215,7 +1215,8 @@
         var base = '<span style="font-family: monospace;visibility:hidden;';
         var ch = $(base + 'width:1ch;overflow: hidden">&nbsp;</span>').appendTo('body');
         var space = $(base + '">&nbsp;</span>').appendTo('body');
-        ch_unit_bug = width(ch) !== width(space);
+        // in FireFox the size of space is fraction larger #579
+        ch_unit_bug = Math.abs(width(ch) - width(space)) > 0.0001;
         ch.remove();
         space.remove();
     });
@@ -2705,6 +2706,15 @@
             return $.terminal.substring(str, start, end);
         }
         // ---------------------------------------------------------------------
+        // :: helper function that check if string is valid emoji formatting
+        // ---------------------------------------------------------------------
+        function is_emoji_formatting(str) {
+            if ($.terminal.is_formatting(str)) {
+                return str.replace(format_parts_re, '$4').match(/^emoji /);
+            }
+            return false;
+        }
+        // ---------------------------------------------------------------------
         // :: Function that displays the command line. Split long lines and
         // :: place cursor in the right place
         // ---------------------------------------------------------------------
@@ -2728,13 +2738,11 @@
                 var position = settings.position;
                 var len = length(string);
                 var prompt = settings.prompt;
-                if (ch_unit_bug) {
-                    cursor.width(char_width);
-                }
                 var c;
                 if (position === len) {
                     before.html(format(string));
-                    cursor.html('<span><span>&nbsp;</span></span>');
+                    c = '&nbsp;';
+                    cursor.html('<span><span>' + c + '</span></span>');
                     after.html('');
                 } else if (position === 0) {
                     before.html('');
@@ -2757,6 +2765,19 @@
                             c_before += c;
                         }
                         after.html(format(substring(string, position + 1), c_before));
+                    }
+                }
+                if (ch_unit_bug) {
+                    if (typeof wcwidth !== 'undefined') {
+                        // handle emoji and wide characters in IE or
+                        // other possible browsers that don't have valid ch unit
+                        var size = strlen(text(c));
+                        if (size === 1 && is_emoji_formatting(c)) {
+                            size = 2;
+                        }
+                        cursor.width(char_width * size);
+                    } else {
+                        cursor.width(char_width);
                     }
                 }
                 cursor.toggleClass('cmd-end-line', cursor_end_line);